feat GoView可视化大屏对接

+++ bootx-services/service-visualization/src/main/java/cn/bootx/visualization/dto/PageResult.java
@@ -16,5 +16,5 @@ import lombok.ToString;
 public class PageResult<T> extends ResResult<T> {
 
     /** 数据条数 */
-    private Integer count;
+    private long count;
 }
